+++ test/lib/Elastica/Test/Multi/SearchTest.php
@@ -575,4 +575,45 @@ class SearchTest extends BaseTest
         $this->assertSame($query2, $resultSets[1]->getQuery());
         $this->assertEquals(6, $resultSets[1]->getTotalHits());
     }
+
+    /**
+     * @group functional
+     */
+    public function testSearchWithSearchOptions()
+    {
+        $type = $this->_createType();
+        $index = $type->getIndex();
+        $client = $index->getClient();
+
+        $multiSearch = new MultiSearch($client);
+
+        $search1 = new Search($client);
+        $search1->addIndex($index)->addType($type);
+        $search1->setOption('terminate_after', '1');
+        $query1 = new Query();
+        $termQuery1 = new Term();
+        $termQuery1->setTerm('username', 'bunny');
+        $query1->setQuery($termQuery1);
+        $query1->setSize(1);
+        $search1->setQuery($query1);
+
+        $multiSearch->addSearch($search1);
+
+        $this->assertCount(1, $multiSearch->getSearches());
+
+        $search2 = new Search($client);
+        $search2->addIndex($index)->addType($type);
+        $query2 = new Query();
+        $termQuery2 = new Term();
+        $termQuery2->setTerm('username', 'bunny');
+        $query2->setQuery($termQuery2);
+        $query2->setSize(3);
+        $search2->setQuery($query2);
+
+        $multiSearch->addSearch($search2);
+        $multiResultSet = $multiSearch->search();
+        $resultSets = $multiResultSet->getResultSets();
+        $this->assertEquals(1, $resultSets[0]->getTotalHits());
+        $this->assertEquals(6, $resultSets[1]->getTotalHits());
+    }
 }